- update naming conventions - move tests out of adviser folder
moved the following to common folder:
- footer views
- robots & security text files
- mixins
- constants
- paginator
- region and utils

Also:
- removed code duplication
- rename 'search' view to 'results' view, as it makes more sense
- add license to cla_search app
- move tests out of adviser folder
- update naming conventions
