+++ fala/common/utils.py
@@ -1,5 +1,4 @@
 # utils.py
-
 import requests
 from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
